<rdar://problem/6752802> LLVM: Input constraint with a matching output constraint of incompatible type To pass LLVM's input/output constraint checker, explicitly drop the most significant bytes by casting to an unsigned int. Note that we want ROTATE to be rotations of 32 bits, not 64.
git-svn-id: http://svn.macosforge.org/repository/mdnsresponder/trunk@6599 214c2c4a-bf3b-4dcf-9390-e4dd3010487d

+++ mDNSCore/DNSDigest.c
@@ -432,12 +432,17 @@ void md5_block_data_order (MD5_CTX *c, const void *p,int num);
    *
    * 					<appro@fy.chalmers.se>
    */
+  /*
+   * LLVM is more strict about compatibility of types between input & output constraints,
+   * but we want these to be rotations of 32 bits, not 64, so we explicitly drop the
+   * most significant bytes by casting to an unsigned int.
+   */
 #  if defined(__i386) || defined(__i386__) || defined(__x86_64) || defined(__x86_64__)
 #   define ROTATE(a,n)	({ register unsigned int ret;	\
 				asm (			\
 				"roll %1,%0"		\
 				: "=r"(ret)		\
-				: "I"(n), "0"(a)	\
+				: "I"(n), "0"((unsigned int)a)	\
 				: "cc");		\
 			   ret;				\
 			})
